refactor: 파일명 변경 - #6
통일성을 위해 파일명을 일부 변경합니다.
- reward-service.js -> reward-model-service.js
- user-service.js -> user-model-service.js
